Testing writing the whole network in Keras
Keras have doesn't really want other tensors as input, you can either input something to the model in the form of Input(tensor=mytensor), but then the input will always have to come from that tensor and not from anywhere else or use Input(shape=(...), name='...') and in that case you can feed a dict with that name but you can only feed a numpy array, not a tensor. For the same reason having tensors as output is not doable and in order to define a loss function one have to input the target tensor as input, make the difference with the net output inside the model and minimize the mse between that difference and a 0 fed from the outside.